page merci + table order

+++ app/Http/Controllers/CheckoutController.php
@@ -2,11 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Order;
+use DateTime;
 use Carbon\Carbon;
 use Gloudemans\Shoppingcart\Facades\Cart;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
 use Stripe\Stripe;
 use Stripe\PaymentIntent;
 
@@ -17,6 +20,8 @@ class CheckoutController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+
     public function index()
     {
 
@@ -25,7 +30,7 @@ class CheckoutController extends Controller
         }
         Stripe::setApiKey('sk_test_51LwRvyGm0FBUZE3XjpQljbBnFygZ9zovVYMTp9qidSYmgfSmWmnqeAZ71hnNt87i6rw3nkvrogcgoXFgoewpDQEV004vopMtT4');
         $intent = PaymentIntent::create([
-            'amount' => round(floatval(Cart::total())),
+            'amount' => round((Cart::total())),
             'currency' => 'eur',
             'metadata' => [
                 'userId' => Auth::user()->id
@@ -56,10 +61,58 @@ class CheckoutController extends Controller
      */
     public function store(Request $request)
     {
-        Cart::destroy();
+       
+       
         $data = $request->json()->all();
+        dump($data);
+       
+        $order = new Order();
 
-        return $data['paymentIntent'];
+        
+        $order->payment_intent_id = $data['paymentIntent']['id'];
+       $order->amount = $data['paymentIntent']['amount'];
+
+        $order->payment_created_at = (new DateTime())->setTimestamp($data['paymentIntent']['created'])->format('Y-m-d H:i:s');
+
+        $products = [];
+        $i = 0;
+
+        foreach(Cart::content() as $product) {
+            $products['product_' . $i][] = $product->model->Name;
+            $products['product_' . $i][] = $product->model->Price;
+            $products['product_' . $i][] = $product->qty;
+            $i++;
+        }
+        
+        $order->products = serialize($products);
+        $order->user_id = Auth::user()->id;
+        $order->save();
+       
+       // Order::create(collect($order)->all());
+
+
+
+
+        
+       
+        if ($data['paymentIntent']['status'] == 'succeeded') {
+            Cart::destroy();
+            Session::flash('success', 'Votre commande a été traitée avec succès.');
+           // return $data['paymentIntent'];
+            return response()->json(['success' => 'Payment Intent Succeeded']);
+        } else {
+            //return $data['paymentIntent'];
+            return response()->json(['error' => 'Payment Intent Not Succeeded']);
+            
+            
+        }
+    }
+
+
+    public function thankYou()
+    {
+        return Session::has('success') ? view('checkout.thankYou') : redirect()->route('home');
+        
     }
 
     /**
